Debug clear_ha

+++ tests/integration/core/clear_ha_el7.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-set -e
+set -xe
 
 if pcs status ; then
     pcs cluster stop --all
@@ -12,12 +12,14 @@ if [ -f /etc/corosync/corosync.conf ]; then
     ring0_iface=$(ip route get "$(sed -ne '/ringnumber: 0/{s///;n;s/.*: //p}' /etc/corosync/corosync.conf)" | sed -ne 's/.* dev \([^ ]*\)  *src.*/\1/p')
 fi
 
-ifconfig "$ring0_iface" 0.0.0.0 down
-
 pcs cluster destroy
 systemctl disable --now pcsd pacemaker corosync
 
-rm -f /etc/sysconfig/network-scripts/ifcfg-"$ring0_iface"
+if [ -n "$ring0_iface" ]; then
+    ifdown "$ring0_iface"
+    rm -f /etc/sysconfig/network-scripts/ifcfg-"$ring0_iface"
+fi
+
 rm -f /etc/corosync/corosync.conf
 rm -f /var/lib/pacemaker/cib/*
 rm -f /var/lib/corosync/*
